Move mdx-namespace to basic
