+++ packages/integrations/mdx/test/fixtures/basic/src/components/namespace/Component.jsx
