+++ packages/integrations/mdx/test/fixtures/basic/src/pages/namespace/object.mdx
@@ -1,0 +1,3 @@
+import * as mod from '../../components/namespace/Component.jsx';
+
+<mod.ns.Component client:load />
