+++ packages/integrations/mdx/test/fixtures/basic/src/pages/namespace/star.mdx
@@ -1,0 +1,3 @@
+import { ns } from '../../components/namespace/Component.jsx';
+
+<ns.Component client:load />
