--- packages/integrations/mdx/test/fixtures/mdx-namespace/package.json
@@ -1,11 +1,0 @@
-{
-  "name": "@test/mdx-namespace",
-  "private": true,
-  "dependencies": {
-    "@astrojs/mdx": "workspace:*",
-    "@astrojs/react": "workspace:*",
-    "astro": "workspace:*",
-    "react": "^18.2.0",
-    "react-dom": "^18.1.0"
-  }
-}
