--- packages/integrations/mdx/test/fixtures/mdx-namespace/src/pages/object.mdx
@@ -1,3 +1,0 @@
-import * as mod from '../components/Component.jsx';
-
-<mod.ns.Component client:load />
