--- packages/integrations/mdx/test/fixtures/mdx-namespace/src/pages/star.mdx
@@ -1,3 +1,0 @@
-import { ns } from '../components/Component.jsx';
-
-<ns.Component client:load />
